[#19] Moved database init functions to their own file so serverless functions are running them each time on startup.
